Add ability to discern between v1 and v2 APIs.
This will be used to help sort out API changes like what was
discovered in #5 (note that this does not implement #5's fix yet,
just lays the groundwork).

**BREAKING CHANGE**
`credhub.New` now returns `(*credhub.Client, error)` instead of
just `*credhub.Client` as was done before, as the client gets the
server version on create.